Not using MtgSDK anymore + any card delimiter
MtgSDK apparently streams their data or something like that, and that was down yesterday. I decided to use a local index of cards instead ; when you start the bot up for the first time, it should downlad the right files.

Also added configuration for card delimiters. The default is now [[card]], but it can be changed to anything.